fix #human to handle ENV['USER'] being nil
if an environment variable doesn't exist, ENV will return nil, not empty
string. it could still be an empty string though (assuming it exists).

change is to call #to_s on assignment (nil.to_s is "") and then
comparing that with empty string.

this way if ENV['USER'] is not set or it's set to empty string, logic
will fall back to "Someone".